Move `Scrollable` to `src` folder, update import paths
